Fix enh-ruby-mode setup
Putting the `add-to-list` call inside the `:config` for enh-ruby-mode
wasn't actually getting it into the right place, but using the `:mode`
keyword in use-package did the trick.

I also added enh-ruby-mode as the ruby-interpreter. I'm not _really_
sure if this is right, but I figured I'd give it a whirl and see what
happens.